+++ backend/app/Service/EmployeeProject/EmployeeProjectService.php
@@ -2,6 +2,7 @@
 
 namespace App\Service\EmployeeProject;
 
+use App\Models\EmployeeAssign;
 use App\Models\EmployeeProject;
 
 class EmployeeProjectService implements EmployeeProjectServiceInterface
@@ -20,7 +21,13 @@ class EmployeeProjectService implements EmployeeProjectServiceInterface
 
     public function delete($id)
     {
-        $data = EmployeeProject::where('id', $id)->first();
+        $data = EmployeeProject::find($id);
+
+        $employeeAssign = EmployeeAssign::where('employee_id',$id)->exists();
+
+        if ($employeeAssign) {
+            return false;
+        }
 
         return $data->delete();
     }
